requirements.txt
#Core Web UI
flask

#Core PDF and image processing
pdfplumber
Pillow

#For interacting with AI services
requests
openai

#For advanced natural language processing (recommended)
spacy
